+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - envs: Allow gain scales larger than one in servo environments
 - envs: Drop Gymnasium environment versioning
 - examples: Allow users to select the Gymnasium environment in MPC example
+- logging: Use an "upkie" logger rather than the global logger
 - spines: Factor common observers and sensors between spines
 - utils: Generalize `abs_bounded_derivative_filter` to vector inputs
 - utils: Generalize `bounded_derivative_filter` to vector inputs